Add missing locations and IDs Add new root AM2RClient.py Moved old AM2RClient.py to world root Update __init__.py
